backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
aiosqlite==0.19.0
sqlalchemy==2.0.25

# AI
dashscope==1.17.0

# Vector Store
chromadb==0.4.22

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Date/Time
python-dateutil==2.8.2